+++ bcdi/postprocessing/postprocessing_utils.py
@@ -37,6 +37,8 @@ def apodize(amp, phase, initial_shape, window_type, debugging=False, **kwargs):
     :param debugging: set to True to see plots
     :type debugging: bool
     :param kwargs:
+
+     - 'cmap': str, name of the colormap
      - for the normal distribution: 'sigma' and 'mu' of the 3d multivariate normal
        distribution, tuples of 3 floats
      - for the Tuckey window: 'alpha' (shape parameter) of the 3d Tukey window,
@@ -50,14 +52,14 @@ def apodize(amp, phase, initial_shape, window_type, debugging=False, **kwargs):
     # check and load kwargs
     valid.valid_kwargs(
         kwargs=kwargs,
-        allowed_kwargs={"sigma", "mu", "alpha", "is_orthogonal"},
+        allowed_kwargs={"cmap", "sigma", "mu", "alpha", "is_orthogonal"},
         name="postprocessing_utils.apodize",
     )
     sigma = kwargs.get("sigma")
     mu = kwargs.get("mu")
     alpha = kwargs.get("alpha")
     is_orthogonal = kwargs.get("is_orthogonal", False)
-
+    cmap = kwargs.get("cmap", "turbo")
     # calculate the diffraction pattern of the reconstructed object
     nb_z, nb_y, nb_x = amp.shape
     nbz, nby, nbx = initial_shape
@@ -73,6 +75,7 @@ def apodize(amp, phase, initial_shape, window_type, debugging=False, **kwargs):
             reciprocal_space=False,
             is_orthogonal=is_orthogonal,
             scale="linear",
+            cmap=cmap,
         )
 
     my_fft = fftshift(fftn(myobj))
@@ -89,6 +92,7 @@ def apodize(amp, phase, initial_shape, window_type, debugging=False, **kwargs):
             reciprocal_space=True,
             is_orthogonal=is_orthogonal,
             scale="log",
+            cmap=cmap,
         )
 
     if window_type == "normal":
@@ -138,6 +142,7 @@ def apodize(amp, phase, initial_shape, window_type, debugging=False, **kwargs):
             reciprocal_space=True,
             is_orthogonal=is_orthogonal,
             scale="log",
+            cmap=cmap,
         )
 
     myobj = ifftn(ifftshift(my_fft))
@@ -152,6 +157,7 @@ def apodize(amp, phase, initial_shape, window_type, debugging=False, **kwargs):
             reciprocal_space=False,
             is_orthogonal=is_orthogonal,
             scale="linear",
+            cmap=cmap,
         )
     myobj = util.crop_pad(
         myobj, (nb_z, nb_y, nb_x)
@@ -571,6 +577,7 @@ def find_bulk(
     width_y=None,
     width_x=None,
     debugging=False,
+    **kwargs,
 ):
     """
     Isolate the inner part of the crystal from the non-physical surface.
@@ -587,9 +594,14 @@ def find_bulk(
      of the initial array
     :param debugging: set to True to see plots
     :type debugging: bool
+    :param kwargs:
+
+     - 'cmap': str, name of the colormap
+
     :return: the support corresponding to the bulk
     """
     valid.valid_ndarray(arrays=amp, ndim=3)
+    cmap = kwargs.get("cmap", "turbo")
 
     nbz, nby, nbx = amp.shape
     max_amp = abs(amp).max()
@@ -622,6 +634,7 @@ def find_bulk(
                 vmin=0,
                 vmax=1,
                 title="Outer matrix",
+                cmap=cmap,
             )
 
         ############################################################################
@@ -659,6 +672,7 @@ def find_bulk(
                     vmin=0,
                     vmax=1,
                     title="Surface matrix",
+                    cmap=cmap,
                 )
 
             # second step: calculate the % of voxels from that layer whose amplitude
@@ -820,17 +834,21 @@ def find_datarange(array, plot_margin=10, amplitude_threshold=0.1, keep_size=Fal
     return zrange, yrange, xrange
 
 
-def flip_reconstruction(obj, debugging=False):
+def flip_reconstruction(obj, debugging=False, **kwargs):
     """
     Calculate the conjugate object giving the same diffracted intensity as 'obj'.
 
     :param obj: 3D reconstructed complex object
     :param debugging: set to True to see plots
     :type debugging: bool
+    :param kwargs:
+
+     - 'cmap': str, name of the colormap
+
     :return: the flipped complex object
     """
     valid.valid_ndarray(arrays=obj, ndim=3)
-
+    cmap = kwargs.get("cmap", "turbo")
     flipped_obj = ifftn(ifftshift(np.conj(fftshift(fftn(obj)))))
     if debugging:
         gu.multislices_plot(
@@ -839,6 +857,7 @@ def flip_reconstruction(obj, debugging=False):
             sum_frames=False,
             plot_colorbar=True,
             title="Initial object",
+            cmap=cmap,
         )
         gu.multislices_plot(
             abs(flipped_obj),
@@ -846,6 +865,7 @@ def flip_reconstruction(obj, debugging=False):
             sum_frames=False,
             plot_colorbar=True,
             title="Flipped object",
+            cmap=cmap,
         )
     return flipped_obj
 
@@ -912,6 +932,8 @@ def get_opticalpath(support, direction, k, voxel_size=None, debugging=False, **k
     :param voxel_size: tuple, actual voxel size in z, y, and x (CXI convention)
     :param debugging: boolena, True to see plots
     :param kwargs:
+
+     - 'cmap': str, name of the colormap
      - width_z: size of the area to plot in z (axis 0), centered on the middle
        of the initial array
      - width_y: size of the area to plot in y (axis 1), centered on the middle
@@ -925,8 +947,11 @@ def get_opticalpath(support, direction, k, voxel_size=None, debugging=False, **k
     # check and load kwargs #
     #########################
     valid.valid_kwargs(
-        kwargs=kwargs, allowed_kwargs={"width_z", "width_y", "width_x"}, name="kwargs"
+        kwargs=kwargs,
+        allowed_kwargs={"cmap", "width_z", "width_y", "width_x"},
+        name="kwargs",
     )
+    cmap = kwargs.get("cmap", "turbo")
     width_z = kwargs.get("width_z")
     valid.valid_item(
         value=width_z,
@@ -1059,6 +1084,7 @@ def get_opticalpath(support, direction, k, voxel_size=None, debugging=False, **k
             title="Support for optical path",
             is_orthogonal=True,
             reciprocal_space=False,
+            cmap=cmap,
         )
 
     ###########################################
@@ -1071,6 +1097,7 @@ def get_opticalpath(support, direction, k, voxel_size=None, debugging=False, **k
         half_width=1,
         title="Optical path",
         debugging=debugging,
+        cmap=cmap,
     )
 
     return path
@@ -1084,6 +1111,7 @@ def get_strain(
     extent_phase=2 * pi,
     method="default",
     debugging=False,
+    **kwargs,
 ):
     """
     Calculate the 3D strain array.
@@ -1101,9 +1129,14 @@ def get_strain(
     :param method: 'default' or 'defect'. If 'defect', will offset the phase
      in a loop and keep the smallest value for the strain (Felix Hofmann's method 2019).
     :param debugging: True to see plots
+    :param kwargs:
+
+     - 'cmap': str, name of the colormap
+
     :return: the strain 3D array
     """
     # check some parameters
+    cmap = kwargs.get("cmap", "turbo")
     valid.valid_ndarray(arrays=phase, ndim=3)
     if reference_axis not in {"x", "y", "z"}:
         raise ValueError("The reference axis should be 'x', 'y' or 'z'")
@@ -1166,6 +1199,7 @@ def get_strain(
                 plot_colorbar=True,
                 is_orthogonal=True,
                 reciprocal_space=False,
+                cmap=cmap,
             )
             gu.multislices_plot(
                 strain,
@@ -1176,6 +1210,7 @@ def get_strain(
                 plot_colorbar=True,
                 is_orthogonal=True,
                 reciprocal_space=False,
+                cmap=cmap,
             )
     return strain
 
@@ -1191,6 +1226,7 @@ def mean_filter(
     vmax=np.nan,
     title="Object",
     debugging=False,
+    **kwargs,
 ):
     """
     Apply a mean filter to an object defined by a support.
@@ -1212,11 +1248,16 @@ def mean_filter(
     :param vmax: real number, higher boundary for the colorbar of the plots
     :param title: str, title for the plots
     :param debugging: bool, True to see plots
+    :param kwargs:
+
+     - 'cmap': str, name of the colormap
+
     :return: averaged array of the same shape as the input array
     """
     #########################
     # check some parameters #
     #########################
+    cmap = kwargs.get("cmap", "turbo")
     valid.valid_ndarray(arrays=(array, support), ndim=3)
     valid.valid_item(half_width, allowed_types=int, min_included=0, name="half_width")
     valid.valid_container(title, container_types=str, name="title")
@@ -1259,6 +1300,7 @@ def mean_filter(
                 vmax=vmax,
                 title=title + " before averaging",
                 plot_colorbar=True,
+                cmap=cmap,
             )
             gu.multislices_plot(
                 support,
@@ -1268,6 +1310,7 @@ def mean_filter(
                 vmin=0,
                 vmax=1,
                 title="Support for averaging",
+                cmap=cmap,
             )
 
         nonzero_pixels = np.argwhere(support != 0)
@@ -1309,6 +1352,7 @@ def mean_filter(
                 vmax=vmax,
                 title=title + " after averaging",
                 plot_colorbar=True,
+                cmap=cmap,
             )
         if counter != 0:
             print("There were", counter, "voxels for which phase could not be averaged")
@@ -1505,6 +1549,8 @@ def remove_offset(
     :param title: string, used in plot title
     :param debugging: True to see plots
     :param kwargs:
+
+     - 'cmap': str, name of the colormap
      - 'reciprocal_space': True if the object is in reciprocal space
      - 'is_orthogonal': True if the data is in an orthonormal frame. Used for defining
        default plot labels.
@@ -1515,9 +1561,10 @@ def remove_offset(
     # check and load kwargs
     valid.valid_kwargs(
         kwargs=kwargs,
-        allowed_kwargs={"reciprocal_space", "is_orthogonal"},
+        allowed_kwargs={"cmap", "reciprocal_space", "is_orthogonal"},
         name="postprocessing_utils.average_obj",
     )
+    cmap = kwargs.get("cmap", "turbo")
     valid.valid_container(
         offset_origin,
         container_types=(tuple, list, np.ndarray),
@@ -1536,6 +1583,7 @@ def remove_offset(
             title=title + " before offset removal",
             reciprocal_space=reciprocal_space,
             is_orthogonal=is_orthogonal,
+            cmap=cmap,
         )
 
     if phase_offset is None:
@@ -1590,6 +1638,7 @@ def remove_offset(
             title=title + " after offset removal",
             reciprocal_space=reciprocal_space,
             is_orthogonal=is_orthogonal,
+            cmap=cmap,
         )
     return array
 
@@ -1606,6 +1655,7 @@ def remove_ramp(
     method="gradient",
     ups_factor=2,
     debugging=False,
+    **kwargs,
 ):
     """
     Remove the linear trend in the ramp using its gradient and a threshold n 3D dataset.
@@ -1628,15 +1678,19 @@ def remove_ramp(
      this value)
     :param debugging: set to True to see plots
     :type debugging: bool
+    :param kwargs:
+
+     - 'cmap': str, name of the colormap
+
     :return: normalized amplitude, detrended phase, ramp along z, ramp along y,
      ramp along x
     """
     valid.valid_ndarray(arrays=(amp, phase), ndim=3)
-
+    cmap = kwargs.get("cmap", "turbo")
     if method == "upsampling":
         nbz, nby, nbx = [mysize * ups_factor for mysize in initial_shape]
         nb_z, nb_y, nb_x = amp.shape
-        myobj = util.crop_pad(amp * np.exp(1j * phase), (nbz, nby, nbx))
+        myobj = util.crop_pad(amp * np.exp(1j * phase), (nbz, nby, nbx), cmap=cmap)
         if debugging:
             plt.figure()
             plt.imshow(np.log10(abs(myobj).sum(axis=0)))
@@ -1648,7 +1702,7 @@ def remove_ramp(
         if debugging:
             plt.figure()
             # plt.imshow(np.log10(abs(my_fft[nbz//2, :, :])))
-            plt.imshow(np.log10(abs(my_fft).sum(axis=0)))
+            plt.imshow(np.log10(abs(my_fft).sum(axis=0)), cmap=cmap)
             plt.title("np.log10(abs(my_fft).sum(axis=0))")
             plt.pause(0.1)
         zcom, ycom, xcom = center_of_mass(abs(my_fft) ** 4)
@@ -1664,7 +1718,7 @@ def remove_ramp(
         gc.collect()
         if debugging:
             plt.figure()
-            plt.imshow(abs(buf2ft).sum(axis=0))
+            plt.imshow(abs(buf2ft).sum(axis=0), cmap=cmap)
             plt.title("abs(buf2ft).sum(axis=0)")
             plt.pause(0.1)
 
@@ -1682,7 +1736,7 @@ def remove_ramp(
         gc.collect()
         if debugging:
             plt.figure()
-            plt.imshow(abs(greg).sum(axis=0))
+            plt.imshow(abs(greg).sum(axis=0), cmap=cmap)
             plt.title("abs(greg).sum(axis=0)")
             plt.pause(0.1)
 
@@ -1693,7 +1747,7 @@ def remove_ramp(
 
         if debugging:
             plt.figure()
-            plt.imshow(np.log10(abs(my_fft).sum(axis=0)))
+            plt.imshow(np.log10(abs(my_fft).sum(axis=0)), cmap=cmap)
             plt.title("centered np.log10(abs(my_fft).sum(axis=0))")
             plt.pause(0.1)
 
@@ -1703,7 +1757,7 @@ def remove_ramp(
         gc.collect()
         if debugging:
             plt.figure()
-            plt.imshow(abs(myobj).sum(axis=0))
+            plt.imshow(abs(myobj).sum(axis=0), cmap=cmap)
             plt.title("centered abs(myobj).sum(axis=0)")
             plt.pause(0.1)
 
@@ -1746,6 +1800,7 @@ def remove_ramp(
             vmin=-threshold_gradient,
             vmax=threshold_gradient,
             title="Phase gradient along Z",
+            cmap=cmap,
         )
         gu.multislices_plot(
             mysupportz,
@@ -1755,6 +1810,7 @@ def remove_ramp(
             vmin=0,
             vmax=1,
             title="Thresholded support along Z",
+            cmap=cmap,
         )
     del mysupportz, mygradz
     gc.collect()
@@ -1779,6 +1835,7 @@ def remove_ramp(
             vmin=-threshold_gradient,
             vmax=threshold_gradient,
             title="Phase gradient along Y",
+            cmap=cmap,
         )
         gu.multislices_plot(
             mysupporty,
@@ -1788,6 +1845,7 @@ def remove_ramp(
             vmin=0,
             vmax=1,
             title="Thresholded support along Y",
+            cmap=cmap,
         )
     del mysupporty, mygrady
     gc.collect()
@@ -1812,6 +1870,7 @@ def remove_ramp(
             vmin=-threshold_gradient,
             vmax=threshold_gradient,
             title="Phase gradient along X",
+            cmap=cmap,
         )
         gu.multislices_plot(
             mysupportx,
@@ -1821,6 +1880,7 @@ def remove_ramp(
             vmin=0,
             vmax=1,
             title="Thresholded support along X",
+            cmap=cmap,
         )
     del mysupportx, mygradx, mysupport
     gc.collect()
@@ -2159,6 +2219,8 @@ def unwrap(obj, support_threshold, seed=0, debugging=True, **kwargs):
      deterministic behavior.
     :param debugging: set to True to see plots
     :param kwargs:
+
+     - 'cmap': str, name of the colormap
      - 'reciprocal_space': True if the object is in reciprocal space
      - 'is_orthogonal': True if the data is in an orthonormal frame.
        Used for defining default plot labels.
@@ -2174,7 +2236,7 @@ def unwrap(obj, support_threshold, seed=0, debugging=True, **kwargs):
     # check and load kwargs
     valid.valid_kwargs(
         kwargs=kwargs,
-        allowed_kwargs={"reciprocal_space", "is_orthogonal"},
+        allowed_kwargs={"cmap", "reciprocal_space", "is_orthogonal"},
         name="postprocessing_utils.average_obj",
     )
     reciprocal_space = kwargs.get("reciprocal_space", False)
@@ -2194,6 +2256,7 @@ def unwrap(obj, support_threshold, seed=0, debugging=True, **kwargs):
             title="Object before unwrapping",
             reciprocal_space=reciprocal_space,
             is_orthogonal=is_orthogonal,
+            cmap=kwargs.get("cmap", "turbo"),
         )
 
     phase_unwrapped = unwrap_phase(phase_wrapped, wrap_around=False, seed=seed).data
@@ -2205,6 +2268,7 @@ def unwrap(obj, support_threshold, seed=0, debugging=True, **kwargs):
             title="Object after unwrapping",
             reciprocal_space=reciprocal_space,
             is_orthogonal=is_orthogonal,
+            cmap=kwargs.get("cmap", "turbo"),
         )
 
     extent_phase = np.ceil(phase_unwrapped.max() - phase_unwrapped.min())