+++ bcdi/utils/parameters.py
@@ -12,6 +12,7 @@ Validation of configuration parameters.
 The validation is performed only on the expected parameters. Other parameters are simply
 discarded.
 """
+import colorcet as cc
 from numbers import Number, Real
 import numpy as np
 import os
@@ -198,6 +199,9 @@ def valid_param(key: str, value: Any) -> Tuple[Any, bool]:
         valid.valid_item(value, allowed_types=int, allow_none=True, name=key)
     elif key == "center_roi_y":
         valid.valid_item(value, allowed_types=int, allow_none=True, name=key)
+    elif key == "colormap":
+        if value not in ["turbo", "custom"] and value not in cc.cm:
+            raise ValueError(f"unknow colormap '{value}'")
     elif key == "comment":
         valid.valid_container(value, container_types=str, name=key)
         if value and not value.startswith("_"):