+++ scripts/postprocessing/bcdi_strain.py
@@ -16,7 +16,7 @@ from bcdi.postprocessing.postprocessing_runner import run
 from bcdi.utils.parser import add_cli_parameters, ConfigParser
 
 here = Path(__file__).parent
-CONFIG_FILE = str(here.parents[1] / "bcdi/examples/config_postprocessing.yml")
+CONFIG_FILE = str(here.parents[1] / "bcdi/examples/S11_config_postprocessing.yml")
 
 helptext = """
 Interpolate the output of the phase retrieval into an orthonormal frame,
@@ -60,6 +60,9 @@ Usage:
     :param comment: string use in filenames when saving
     :param debug: e.g. False
      True to see plots
+    :param colormap: e.g. "turbo"
+     "turbo", "custom" or colormap defined in the colorcet package, see
+     https://colorcet.holoviz.org/
     :param reconstruction_file: e.g. "modes.h5"
      path to a reconstruction file, to avoid opening a pop-up window