+++ tests/utils/test_parser.py
@@ -12,7 +12,7 @@ from bcdi.utils.parser import ConfigParser
 from tests.config import run_tests
 
 here = Path(__file__).parent
-CONFIG = str(here.parents[1] / "bcdi/examples/config_postprocessing.yml")
+CONFIG = str(here.parents[1] / "bcdi/examples/S11_config_postprocessing.yml")
 
 
 class TestConfigParser(unittest.TestCase):
@@ -23,7 +23,11 @@ class TestConfigParser(unittest.TestCase):
     """
 
     def setUp(self) -> None:
-        self.command_line_args = {"scan": 999999999, "root_folder": str(here)}
+        self.command_line_args = {
+            "data_dir": str(here),
+            "scan": 999999999,
+            "root_folder": str(here),
+        }
         self.parser = ConfigParser(CONFIG, self.command_line_args)
 
     def test_init_file_path(self):
@@ -76,7 +80,12 @@ class TestConfigParser(unittest.TestCase):
 
     def test_load_arguments_cl_params_flip(self):
         self.parser = ConfigParser(
-            CONFIG, {"flip_reconstruction": "False", "root_folder": str(here)}
+            CONFIG,
+            {
+                "data_dir": str(here),
+                "flip_reconstruction": "False",
+                "root_folder": str(here),
+            },
         )
         # "flip_reconstruction" is also key in CONFIG, which means that the overriding
         # by the optional --flip_reconstruction argument from the command line works as
@@ -86,7 +95,12 @@ class TestConfigParser(unittest.TestCase):
 
     def test_load_arguments_cl_params_flip_no_bool(self):
         self.parser = ConfigParser(
-            CONFIG, {"flip_reconstruction": "weirdstring", "root_folder": str(here)}
+            CONFIG,
+            {
+                "data_dir": str(here),
+                "flip_reconstruction": "weirdstring",
+                "root_folder": str(here),
+            },
         )
         with self.assertRaises(TypeError):
             self.parser.load_arguments()